+++ Backend/src/main/java/com/swp391/gr3/ev_management/repository/ConnectorTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ConnectorTypeRepository extends JpaRepository<ConnectorType, Integer> {
+public interface ConnectorTypeRepository extends JpaRepository<ConnectorType, Long> {
     boolean existsByCode(String code);
     ConnectorType findByCode(String code);
     List<ConnectorType> findDistinctByChargingPoints_Station_StationId(Long stationId);
